+++ changelog.md
@@ -1,5 +1,5 @@
 # Cassiopeia Changelog
-#### 2.0.239.20
+#### 2.0.239.30
 - Cassiopeia autodetects the system language in order to set itself.
 - Fixed context menu when not selecting any album [FIX]
 - Fixed context menu when view is on CD or Vinyl mode [FIX]
@@ -8,6 +8,8 @@
 - Removed localization if album doesn't have a location folder [FIX]
 - Fixed lyrics icon not removing when dropping the selection [FIX]
 - Fixed creating new album from a directory [FIX]
+- Fixed the location of the sound folder in EditAlbum [FIX]
+- Fixed the linking of sound files and album [FIX]
 
 #### 2.0.238.0
 - Added 3 new icons. Add CD, Add vinyl and a search icon to illustrate the textbox.